Add ArrayBuffer support in k6/ws module
I experimented with overriding the first argument passed to the
'message' handler to be a String object with a `buffer` data property,
but it doesn't expose the string primitive properties (.length, etc.) so
it would be a breaking change.

Passing the ArrayBuffer as the second argument is backwards compatible,
and hopefully doesn't cause a doubling of memory usage (I'm expecting it
to be garbage collected if the argument isn't used). We could detect the
number of arguments defined on the handler and only pass the ArrayBuffer
if expected, but this would be awkward to implement.

Part of #1020