Rewrite statsd collector as output
Also deprecate the datadog one where instead now the
statsd output should be used with configuring
`K6_STATSD_TAG_FORMAT=datadog` . The alternative values are `newrelic`,
`statsdv2` and `none`. All but the last one mean are the same format as
that is the only format the currently library support. While the last
one `none` means no tags which is the default and is what statsd used to
support until v0.9.0.

Additionally, a new K6_STATSD_TAG_BLOCKLIST was added analogous to the
K6_DATADOG_TAG_BLACKLIST.

Co-authored-by: na-- <n@andreev.sh>